Added @shutdown for server shutdown (with player save) Added @info for global notice (the blue message)
Corrected some map names, that the @tele command is working right

Fixed a bug with the SlayerWeapon (now @d [SLAYERID] does work)

git-svn-id: http://reunion.googlecode.com/svn/trunk@242 d5b2e400-7aba-11de-9499-4f8beaa274e3

+++ jreunion/src/com/googlecode/reunion/jreunion/game/items/equipment/SlayerWeapon.java
@@ -33,7 +33,7 @@ public class SlayerWeapon extends SpecialWeapon {
 		} else {
 			if (item.checkMembers(new String[] { "MemoryDmg" })) {
 				// use member from file
-				setMemoryDmg(Integer.parseInt(item.getMemberValue("MemoryDmg")));
+				setMemoryDmg(Float.parseFloat(item.getMemberValue("MemoryDmg")));
 			} else {
 				// use default
 				setMemoryDmg(0);
